move int functions to intfunc
