+++ sympy/core/power.py
@@ -9,6 +9,7 @@ from .expr import Expr
 from .evalf import PrecisionExhausted
 from .function import (expand_complex, expand_multinomial,
     expand_mul, _mexpand, PoleError)
+from .intfunc import isqrt, integer_log, integer_nthroot
 from .logic import fuzzy_bool, fuzzy_not, fuzzy_and, fuzzy_or
 from .parameters import global_parameters
 from .relational import is_gt, is_lt
@@ -18,140 +19,6 @@ from sympy.utilities.iterables import sift
 from sympy.utilities.exceptions import sympy_deprecation_warning
 from sympy.utilities.misc import as_int
 from sympy.multipledispatch import Dispatcher
-
-
-def isqrt(n):
-    r""" Return the largest integer less than or equal to `\sqrt{n}`.
-
-    Parameters
-    ==========
-
-    n : non-negative integer
-
-    Returns
-    =======
-
-    int : `\left\lfloor\sqrt{n}\right\rfloor`
-
-    Raises
-    ======
-
-    ValueError
-        If n is negative.
-    TypeError
-        If n is of a type that cannot be compared to ``int``.
-        Therefore, a TypeError is raised for ``str``, but not for ``float``.
-
-    Examples
-    ========
-
-    >>> from sympy.core.power import isqrt
-    >>> isqrt(0)
-    0
-    >>> isqrt(9)
-    3
-    >>> isqrt(10)
-    3
-    >>> isqrt("30")
-    Traceback (most recent call last):
-        ...
-    TypeError: '<' not supported between instances of 'str' and 'int'
-    >>> from sympy.core.numbers import Rational
-    >>> isqrt(Rational(-1, 2))
-    Traceback (most recent call last):
-        ...
-    ValueError: n must be nonnegative
-
-    """
-    if n < 0:
-        raise ValueError("n must be nonnegative")
-    return int(sqrt(int(n)))
-
-
-def integer_nthroot(y, n):
-    """
-    Return a tuple containing x = floor(y**(1/n))
-    and a boolean indicating whether the result is exact (that is,
-    whether x**n == y).
-
-    Examples
-    ========
-
-    >>> from sympy import integer_nthroot
-    >>> integer_nthroot(16, 2)
-    (4, True)
-    >>> integer_nthroot(26, 2)
-    (5, False)
-
-    To simply determine if a number is a perfect square, the is_square
-    function should be used:
-
-    >>> from sympy.ntheory.primetest import is_square
-    >>> is_square(26)
-    False
-
-    See Also
-    ========
-    sympy.ntheory.primetest.is_square
-    integer_log
-    """
-    x, b = iroot(as_int(y), as_int(n))
-    return int(x), b
-
-
-def integer_log(y, x):
-    r"""
-    Returns ``(e, bool)`` where e is the largest nonnegative integer
-    such that :math:`|y| \geq |x^e|` and ``bool`` is True if $y = x^e$.
-
-    Examples
-    ========
-
-    >>> from sympy import integer_log
-    >>> integer_log(125, 5)
-    (3, True)
-    >>> integer_log(17, 9)
-    (1, False)
-    >>> integer_log(4, -2)
-    (2, True)
-    >>> integer_log(-125,-5)
-    (3, True)
-
-    See Also
-    ========
-    integer_nthroot
-    sympy.ntheory.primetest.is_square
-    sympy.ntheory.factor_.multiplicity
-    sympy.ntheory.factor_.perfect_power
-    """
-    if x == 1:
-        raise ValueError('x cannot take value as 1')
-    if y == 0:
-        raise ValueError('y cannot take value as 0')
-
-    if x in (-2, 2):
-        x = int(x)
-        y = as_int(y)
-        e = y.bit_length() - 1
-        return e, x**e == y
-    if x < 0:
-        n, b = integer_log(y if y > 0 else -y, -x)
-        return n, b and bool(n % 2 if y < 0 else not n % 2)
-
-    x = as_int(x)
-    y = as_int(y)
-    r = e = 0
-    while y >= x:
-        d = x
-        m = 1
-        while y >= d:
-            y, rem = divmod(y, d)
-            r = r or rem
-            e += m
-            if y > d:
-                d *= d
-                m *= 2
-    return e, r == 0 and y == 1
 
 
 class Pow(Expr):
