+++ sympy/ntheory/factor_.py
@@ -13,7 +13,8 @@ from sympy.core.function import Function
 from sympy.core.logic import fuzzy_and
 from sympy.core.mul import Mul
 from sympy.core.numbers import Rational, Integer
-from sympy.core.power import Pow, integer_log
+from sympy.core.intfunc import trailing, integer_nthroot, integer_log
+from sympy.core.power import Pow
 from sympy.core.random import _randint
 from sympy.core.singleton import S
 from sympy.external.gmpy import SYMPY_INTS, gcd, lcm, sqrt as isqrt, sqrtrem, iroot
@@ -56,11 +57,6 @@ def _isperfect(n):
             PERFECT.append(t*(2*t - 1))
             j += 1
     return n in PERFECT
-
-
-small_trailing = [0] * 256
-for j in range(1,8):
-    small_trailing[1<<j::1<<(j+1)] = [j] * (1<<(7-j))
 
 
 def smoothness(n):
@@ -193,57 +189,6 @@ def smoothness_p(n, m=-1, power=0, visual=None):
         dat.insert(2, m)
         lines.append('p**i=%i**%i has p%+i B=%i, B-pow=%i' % tuple(dat))
     return '\n'.join(lines)
-
-
-def trailing(n):
-    """Count the number of trailing zero digits in the binary
-    representation of n, i.e. determine the largest power of 2
-    that divides n.
-
-    Examples
-    ========
-
-    >>> from sympy import trailing
-    >>> trailing(128)
-    7
-    >>> trailing(63)
-    0
-
-    See Also
-    ========
-
-    multiplicity
-
-    """
-    n = abs(int(n))
-    if not n:
-        return 0
-    low_byte = n & 0xff
-    if low_byte:
-        return small_trailing[low_byte]
-
-    t = 8
-    n >>= 8
-    # 2**m is quick for z up through 2**30
-    z = n.bit_length() - 1
-    if n == 1 << z:
-        return z + t
-
-    if z < 300:
-        # fixed 8-byte reduction
-        while not n & 0xff:
-            n >>= 8
-            t += 8
-    else:
-        # binary reduction important when there might be a large
-        # number of trailing 0s
-        p = z >> 1
-        while not n & 0xff:
-            while n & ((1 << p) - 1):
-                p >>= 1
-            n >>= p
-            t += p
-    return t + small_trailing[n & 0xff]
 
 
 def multiplicity(p, n):
@@ -1426,11 +1371,11 @@ def factorint(n, limit=None, use_trial=True, use_rho=True, use_pm1=True,
             if verbose:
                 print(complete_msg)
             return factors
-        #Use subexponential algorithms if use_ecm
-        #Use pollard algorithms for finding small factors for 3 iterations
-        #if after small factors the number of digits of n is >= 20 then use ecm
+        # Use subexponential algorithms if use_ecm
+        # Use pollard algorithms for finding small factors for 3 iterations
+        # if after small factors the number of digits of n >= 25 then use ecm
         iteration += 1
-        if use_ecm and iteration >= 3 and ndigits(n) >= 25:
+        if use_ecm and iteration >= 3 and ndigits(n) >= 24:
             break
         low, high = high, high*2
     B1 = 10000
